Town Generator
A generator for a variety of types and sizes of towns, along with the prominent buildings that you would find in that town. The primary 2 functions are Town Size and Town Type.  Everything else is randomized based on one or both of those variables, with sub-tables offering variety so we don't end up with lots of duplicate towns.